requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
asyncpg==0.30.0
sqlalchemy==2.0.41
pillow>=9.0.0
pydantic==2.11.5
pydantic-settings==2.2.1
python-dotenv==1.1.0
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
fastapi-limiter==0.1.6
redis==6.2.0
cloudinary==1.44.0
aiofiles==24.1.0
alembic==1.16.1
email-validator==2.1.1
PyJWT==2.10.1
email-validator>=1.3.0
